internal: improved lexer performance
The lexer and parser are merged into one file, everything is moved to the internal folder.
Added comments.
