+++ internal/styles.go
@@ -1,10 +1,11 @@
-package styles
+package internal
 
 import (
 	"github.com/gookit/color"
 	"github.com/muesli/termenv"
 )
 
+// Bold provides a bold style.
 func Bold(f func(text string) string) func(text string) string {
 	return func(text string) string {
 		t := f(text)
@@ -13,6 +14,7 @@ func Bold(f func(text string) string) func(text string) string {
 	}
 }
 
+// Italic provides a italic style.
 func Italic(f func(text string) string) func(text string) string {
 	return func(text string) string {
 		t := f(text)
@@ -21,6 +23,7 @@ func Italic(f func(text string) string) func(text string) string {
 	}
 }
 
+// CrossOut provides a crossout style.
 func CrossOut(f func(text string) string) func(text string) string {
 	return func(text string) string {
 		t := f(text)
@@ -29,6 +32,7 @@ func CrossOut(f func(text string) string) func(text string) string {
 	}
 }
 
+// Underline provides a underline style.
 func Underline(f func(text string) string) func(text string) string {
 	return func(text string) string {
 		t := f(text)
@@ -37,6 +41,7 @@ func Underline(f func(text string) string) func(text string) string {
 	}
 }
 
+// Overline provides a overline style.
 func Overline(f func(text string) string) func(text string) string {
 	return func(text string) string {
 		t := f(text)
